requirements.txt
streamlit
google-generativeai
rdflib